+++ src/containers/Form/Skills/index.js
@@ -15,7 +15,7 @@ const CreateSkillsForm = compose(
   reduxForm({
     onSubmit: formSubmission,
     validate: formValidation,
-    destroyOnUnmount: false
+    destroyOnUnmount: true
   })
 )(SkillsForm);