+++ src/containers/Form/Users/index.js
@@ -22,7 +22,7 @@ const CreateUsersForm = compose(
   reduxForm({
     onSubmit: formSubmission,
     validate: formValidation,
-    destroyOnUnmount: false
+    destroyOnUnmount: true
   })
 )(UsersForm);