+++ src/containers/SidePanelDetails/Roles/layout.js
@@ -11,8 +11,6 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-
-import { DetailWrapper } from '../../../components/DetailWrapper';
 import { Detail, SidePanelTable, DetailHeader, DetailsPanelAlert } from 'cx-ui-components';
 
 const Wrapper = styled.div`
@@ -20,15 +18,10 @@ const Wrapper = styled.div`
   flex-direction: column;
   height: calc(100vh - 180px);
 `;
-const WrappedDetailHeader = styled(DetailHeader)`
-  margin-left: 35px;
-`;
 
 export function RolesDetailsPanel({
   children,
   userHasUpdatePermission,
-  id,
-  className,
   tableItems,
   tableFields,
   inherited,
@@ -38,42 +31,35 @@ export function RolesDetailsPanel({
   item: { name, description }
 }) {
   return (
-    <Wrapper id={id} className={className}>
-      <DetailWrapper open={true}>
-        <WrappedDetailHeader text="Details" />
-        {inherited && <DetailsPanelAlert text="This role is inherited and cannot be edited" />}
-        {!inherited && userHasUpdatePermission ? (
-          children
-        ) : (
-          <Fragment>
-            <Detail label="Name" value={name} />
-            <Detail label="Description" value={description} />
-          </Fragment>
-        )}
-      </DetailWrapper>
+    <Wrapper id="dtpanel-roles">
+      {inherited && <DetailsPanelAlert text="This role is inherited and cannot be edited" />}
+      {!inherited && userHasUpdatePermission ? (
+        children
+      ) : (
+        <Fragment>
+          <Detail label="Name" value={name} />
+          <Detail label="Description" value={description} />
+        </Fragment>
+      )}
 
-      <DetailWrapper open={true}>
-        <WrappedDetailHeader
-          userHasUpdatePermission={userHasUpdatePermission}
-          text={`${listSize} Permission(s)`}
-          onActionButtonClick={() => setSelectedSubEntityId('addItemToList')}
-          inherited={inherited}
-        />
-        <SidePanelTable
-          userHasUpdatePermission={userHasUpdatePermission}
-          deleteSubEntity={removeListItem}
-          items={tableItems}
-          inherited={inherited}
-          fields={tableFields}
-        />
-      </DetailWrapper>
+      <DetailHeader
+        userHasUpdatePermission={userHasUpdatePermission}
+        text={`${listSize} Permission(s)`}
+        onActionButtonClick={() => setSelectedSubEntityId('addItemToList')}
+        inherited={inherited}
+      />
+      <SidePanelTable
+        userHasUpdatePermission={userHasUpdatePermission}
+        deleteSubEntity={removeListItem}
+        items={tableItems}
+        inherited={inherited}
+        fields={tableFields}
+      />
     </Wrapper>
   );
 }
 
 RolesDetailsPanel.propTypes = {
-  id: PropTypes.string,
-  className: PropTypes.string,
   item: PropTypes.shape({
     name: PropTypes.string,
     description: PropTypes.string