+++ CHANGELOG.md
@@ -13,9 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Skills
 - Groups
 - Users
+- Data Access Control
 
 ## [Beta User Acceptance Testing]
 - Interaction Monitoring
+
+## [0.37.0] - 2018-11-05
+### Added
+- CXV1-14795 - Create Access Control Page with Table Config.
+- CXV1-14812 - Create Access Controlled Reports Side Panel Form.
+
+### Fixed
+- CXV1-15428 - Fixed bug where input values were not cleared when creating new outbound identifier
 
 ## [0.36.0] - 2018-11-05
 ### Added